requirements.txt
numpy==1.24.2
kepler.py==0.0.7
pyquaternion==0.9.9
pymap3d==1.7.15
pytest==7.2.1
